+++ Makefile
@@ -1,24 +1,21 @@
-.POSIX:
-PREFIX = ~/.local
-CONFIG = ~/.config
-PIPVER = `command -v pip3 || command -v pip`
+PREFIX 			= ~/.local
+CONFIG 			= ~/.config
+PIPVER 			= `command -v pip3 || command -v pip`
+CONF_FILE 	= $(XDG_CONFIG_HOME)/chameleon/config.yml
+CURRENT_DIR = $(PWD)
 
 install:
-	mkdir -p $(DESTDIR)$(PREFIX)/bin
-	cp -f chameleon.py $(DESTDIR)$(PREFIX)/bin/chameleon.py
-	@echo "chameleon.py has been installed to $(DESTDIR)$(PREFIX)/bin/chameleon.py"
-	mkdir -p $(DESTDIR)$(CONFIG)/chameleon
-	./setup-config >> $(DESTDIR)$(CONFIG)/chameleon/config.yaml
-	@echo "config file for chameleon.py created in $(DESTDIR)$(PREFIX)/config/config.yaml"
+	mkdir -p $(PREFIX)/bin
+	cp -f chameleon.py $(PREFIX)/bin/chameleon.py
+	@echo "chameleon.py has been installed to $(PREFIX)/bin/chameleon.py"
+	mkdir -p $(CONFIG)/chameleon
+	-mv $(CONF_FILE) $(CONF_FILE)-backup 2>/dev/null
+	-cp -r $(CURRENT_DIR)/config-template.yml $(CONF_FILE)
 	@echo "installing $(PIPVER) dependencies"
 	@$(PIPVER) install whichcraft || echo "dependencies couldn't be installed install pip and rerun"
 	@$(PIPVER) install pywal || echo "dependencies couldn't be installed install pip and rerun"
-	@echo "backing up old config file"
-	mv config.yaml config.yaml.bak
 uninstall:
-	rm -f $(DESTDIR)$(PREFIX)/bin/chameleon.py
-	@echo "removed $(DESTDIR)$(PREFIX)/bin/chameleon.py"
-	rm -rf $(DESTDIR)$(CONFIG)/chameleon
-	@echo "removed $(DESTDIR)$(CONFIG)/chameleon"
-	$(PIPVER) uninstall whichcraft
-	.PHONY: install uninstall pipversion
+	-rm -f $(PREFIX)/bin/chameleon.py
+	@echo "removed $(PREFIX)/bin/chameleon.py"
+	-rm -rf $(CONFIG)/chameleon
+	@echo "removed $(CONFIG)/chameleon"
