--- setup-config
@@ -1,64 +1,0 @@
-#!/usr/bin/bash
-
-cat <<EOF
-oomoxicons:
-  command: "oomox-archdroid-icons-cli"
-  themepath: "/opt/oomox/scripted_colors/xresources/xresources-reverse"
-
-oomoxgtk:
-  themepath: "/opt/oomox/scripted_colors/xresources/xresources-reverse"
-
-oomoxspotify:
-  spotifypath: "/opt/spotify/Apps"
-  enabled: True
-
-slickpywal:
-  path: ""
-  config_path: ""
-  executable: ""
-  arguments: ""
-
-pywalneopixels:
-  path: ""
-  config_path: ""
-  executable: ""
-  arguments: ""
-
-razercli:
-  path: ""
-  config_path: ""
-  executable: ""
-  arguments: ""
-
-telegrampalletegen:
-  path: ""
-  config_path: ""
-  executable: ""
-  arguments: ""
-
-pywalfox:
-  path: ""
-  config_path: ""
-  executable: ""
-  arguments: ""
-
-wal-discord:
-  path: "/home/singularis/.local/bin"
-  executable: "wal-discord"
-  arguments: "beautifuldiscord --css ~/.cache/wal-discord/style.css"
-
-xmenu:
-  path: "/home/singularis/.config/xmenu"
-  arguments: "sudo make clean install"
-
-starttree:
-  path: "/home/singularis/.local/share/Singularis/third-party-tools/start-tree"
-
-gnuplot:
-  path: "/home/singularis/.local/share/Singularis/third-party-tools/gnuplot-pywal"
-
-xfce4-terminal:
-  path: "/home/singularis/.local/share/Singularis/third-party-tools/chameleon/scripts"
-  config_path: "/home/singularis/.config/xfce4/terminal/terminalrc"
-  executable: "./xfce4-terminal.sh"
-EOF
